+++ src/js/const/atomic-swap.const.js
@@ -1,0 +1,5 @@
+export const ATOMIC_SWAP_REQUEST_TYPES = {
+  createOffer: 'create-offer',
+  createPaymentMethod: 'create-payment-method',
+  createBuyRequest: 'create-buy-request',
+}