+++ src/vue/forms/PayForm.vue
@@ -1,22 +1,31 @@
 <template>
   <div>
-    <div
-      class="atomic-swap__asset-description"
-      v-if="assetByCode(atomicSwapAsk.baseAssetCode).description"
-    >
-      <p
-        class="atomic-swap__asset-description-lbl"
-      >
-        {{ 'buy-atomic-swap-form.asset-description-lbl' | globalize }}:
-      </p>
-      <p>{{ assetByCode(atomicSwapAsk.baseAssetCode).description }}</p>
-    </div>
-    <buy-atomic-swap-form
+    <form
       v-if="!isAtomicSwapBidCreated"
-      :atomic-swap-ask="atomicSwapAsk"
-      :is-disabled="isDisabled"
-      @submitted="handleAtomicSwapFormSubmitted"
-    />
+      novalidate
+      class="pay-form app__form"
+    >
+      <div class="app__form-row">
+        <div class="app__form-field">
+          <input-field
+            v-model="form.email"
+            @blur="touchField('form.email')"
+            name="pay-email"
+            :disabled="isLoggedIn"
+            :label="'pay-form.email-lbl' | globalize"
+            :error-message="getFieldErrorMessage('form.email')"
+          />
+        </div>
+      </div>
+
+      <buy-atomic-swap-form
+        class="pay-form__atomic-swap-bid"
+        @submitted="submit"
+        :is-disabled="isDisabled"
+        :atomic-swap-ask="atomicSwapAsk"
+      />
+    </form>
+
     <address-viewer
       v-else
       :asset-code="form.quoteAssetCode"
@@ -28,37 +37,29 @@
 </template>
 
 <script>
-import BuyAtomicSwapForm from '@/vue/forms/BuyAtomicSwapForm'
 import FormMixin from '@/vue/mixins/form.mixin'
+import BuyAtomicSwapForm from '@/vue/forms/BuyAtomicSwapForm'
 import AddressViewer from '@/vue/common/address-viewer'
 import AtomicSwapBidMixin from '@/vue/mixins/atomic-swap-bid.mixin'
+import { required, email } from '@validators'
 import { AtomicSwapAskRecord } from '@/js/records/entities/atomic-swap-ask.record'
+import { ATOMIC_SWAP_BID_TYPES } from '@/js/const/atomic-swap-bid-types.const'
 import { ErrorHandler } from '@/js/helpers/error-handler'
 import { vuexTypes } from '@/vuex'
 import { mapGetters } from 'vuex'
 
-import { ATOMIC_SWAP_BID_TYPES } from '@/js/const/atomic-swap-bid-types.const'
-
-const EVENTS = {
-  updateList: 'update-list',
-}
-
 export default {
-  name: 'atomic-swap',
+  name: 'pay-form',
   components: {
     BuyAtomicSwapForm,
     AddressViewer,
   },
   mixins: [FormMixin, AtomicSwapBidMixin],
-  props: {
-    atomicSwapAsk: {
-      type: AtomicSwapAskRecord,
-      required: true,
-    },
-  },
+  props: { atomicSwapAsk: { type: AtomicSwapAskRecord, required: true } },
   data () {
     return {
       form: {
+        email: '',
         amount: '',
         quoteAssetCode: '',
         paymentMethodId: '',
@@ -71,17 +72,34 @@ export default {
       },
     }
   },
+
+  validations () {
+    return {
+      form: {
+        email: {
+          required,
+          email,
+        },
+      },
+    }
+  },
+
   computed: {
     ...mapGetters([
-      vuexTypes.assetByCode,
+      vuexTypes.walletEmail,
     ]),
 
     isAtomicSwapBidCreated () {
       return Boolean(this.atomicSwapBidDetails.address)
     },
   },
+
+  created () {
+    this.form.email = this.walletEmail
+  },
+
   methods: {
-    async handleAtomicSwapFormSubmitted (form) {
+    async submit (form) {
       if (!this.isFormValid()) return
       Object.assign(this.form, form)
 
@@ -91,13 +109,13 @@ export default {
         const atomicSwapBid = await this.createAtomicSwapBidOperation(
           this.form.amount,
           this.form.paymentMethodId,
-          this.atomicSwapAsk.id
+          this.atomicSwapAsk.id,
+          this.form.email
         )
         if (atomicSwapBid.type === ATOMIC_SWAP_BID_TYPES.redirect) {
           window.location.href = atomicSwapBid.payUrl
         } else {
           this.atomicSwapBidDetails = atomicSwapBid
-          this.$emit(EVENTS.updateList)
         }
       } catch (e) {
         ErrorHandler.process(e)
@@ -109,14 +127,9 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-  @import '~@/scss/variables';
+  @import '~@/vue/forms/app-form';
 
-  .atomic-swap__asset-description {
-    margin-bottom: 2.4rem;
-  }
-
-  .atomic-swap__asset-description-lbl {
-    font-size: 1.2rem;
-    color: $col-primary-inactive;
+  .pay-form__atomic-swap-bid {
+    margin-top: 2.4rem;
   }
 </style>