+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 ## [Unreleased]
+#### Added
+- Pay page
+
 #### Changed
 - Show all base assets in set default quote asset select
 - Renamed direct sell to offering
+- Used new atomic swap requests
 
 #### Fixed
 - A bug when don't show an error message when field was touched